+++ src/tokens/tokens.service.ts
@@ -32,18 +32,13 @@ import ERC721NoDataABI from '../abi/ERC721NoData.json';
 import ERC721WithDataABI from '../abi/ERC721WithData.json';
 import TokenFactoryABI from '../abi/TokenFactory.json';
 import IERC165ABI from '../abi/IERC165.json';
-import { Event, EventStream, EventStreamReply } from '../event-stream/event-stream.interfaces';
+import { EventStream, EventStreamReply } from '../event-stream/event-stream.interfaces';
 import { EventStreamService } from '../event-stream/event-stream.service';
 import { EventStreamProxyGateway } from '../eventstream-proxy/eventstream-proxy.gateway';
-import { EventListener, EventProcessor } from '../eventstream-proxy/eventstream-proxy.interfaces';
 import { basicAuth } from '../utils';
-import { WebSocketMessage } from '../websocket-events/websocket-events.base';
 import { FFRequestIDHeader } from '../request-context/constants';
 import { Context, newContext } from '../request-context/request-context.decorator';
 import {
-  ERC20ApprovalEvent,
-  ERC721ApprovalEvent,
-  ApprovalForAllEvent,
   AsyncResponse,
   ContractSchema,
   EthConnectAsyncResponse,
@@ -52,24 +47,17 @@ import {
   IPoolLocator,
   IValidPoolLocator,
   TokenApproval,
-  TokenApprovalEvent,
   TokenBurn,
-  TokenBurnEvent,
-  TokenPoolCreationEvent,
   TokenMint,
-  TokenMintEvent,
   TokenPool,
   TokenPoolActivate,
   TokenPoolConfig,
   TokenPoolEvent,
   TokenTransfer,
-  TokenTransferEvent,
   TokenType,
-  TransferEvent,
   TokenPoolEventInfo,
 } from './tokens.interfaces';
 import {
-  decodeHex,
   encodeHex,
   getTokenSchema,
   packPoolLocator,
@@ -78,6 +66,7 @@ import {
   unpackSubscriptionName,
   validatePoolLocator,
 } from './tokens.util';
+import { TokenListener } from './tokens.listener';
 
 const ERC20WithDataIID = '0xaefdad0f';
 const ERC721WithDataIID = '0xb2429c12';
@@ -184,16 +173,12 @@ abiEventMap.set('ERC721WithData', {
 
 const tokenCreateMethod = 'create';
 const tokenCreateEvent = 'TokenPoolCreation';
-const tokenCreateEventSignature = 'TokenPoolCreation(address,string,string,bool,bytes)';
 
 const UINT256_MAX = BigInt(2) ** BigInt(256) - BigInt(1);
 const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000';
 
 const sendTransactionHeader = 'SendTransaction';
 const queryHeader = 'Query';
-const transferEventSignature = 'Transfer(address,address,uint256)';
-const approvalEventSignature = 'Approval(address,address,uint256)';
-const approvalForAllEventSignature = 'ApprovalForAll(address,address,bool)';
 
 type ContractSchemaStrings = keyof typeof ContractSchema;
 
@@ -878,313 +863,5 @@ export class TokensService {
       throw new NotFoundException();
     }
     return response.data;
-  }
-}
-
-class TokenListener implements EventListener {
-  private readonly logger = new Logger(TokenListener.name);
-
-  constructor(private readonly service: TokensService) {}
-
-  async onEvent(subName: string, event: Event, process: EventProcessor) {
-    const signature = this.trimEventSignature(event.signature);
-    switch (signature) {
-      case tokenCreateEventSignature:
-        process(await this.transformTokenPoolCreationEvent(subName, event));
-        break;
-      case transferEventSignature:
-        process(await this.transformTransferEvent(subName, event));
-        break;
-      case approvalEventSignature:
-        process(this.transformApprovalEvent(subName, event));
-        break;
-      case approvalForAllEventSignature:
-        process(this.transformApprovalForAllEvent(subName, event));
-        break;
-      default:
-        this.logger.error(`Unknown event signature: ${event.signature}`);
-    }
-  }
-
-  private async getTokenUri(
-    ctx: Context,
-    tokenIdx: string,
-    contractAddress: string,
-  ): Promise<string> {
-    const abiMethods = abiSchemaMap.get('ERC721WithData');
-    if (abiMethods === undefined) {
-      // should not happen
-      return '';
-    }
-
-    const methodABI = abiMethods.find(method => method.name === 'tokenURI');
-    try {
-      const response = await this.service.query(ctx, contractAddress, methodABI, [tokenIdx]);
-      return response.output as string;
-    } catch (e) {
-      this.logger.log(`Burned tokens do not have a URI: ${e}`);
-      return '';
-    }
-  }
-
-  /**
-   * Generate an event ID in the recognized FireFly format for Ethereum
-   * (zero-padded block number, transaction index, and log index)
-   */
-  private formatBlockchainEventId(event: Event) {
-    const blockNumber = event.blockNumber ?? '0';
-    const txIndex = BigInt(event.transactionIndex).toString(10);
-    const logIndex = event.logIndex ?? '0';
-    return [
-      blockNumber.padStart(12, '0'),
-      txIndex.padStart(6, '0'),
-      logIndex.padStart(6, '0'),
-    ].join('/');
-  }
-
-  private stripParamsFromSignature(signature: string) {
-    return signature.substring(0, signature.indexOf('('));
-  }
-
-  private trimEventSignature(signature: string) {
-    const firstColon = signature.indexOf(':');
-    if (firstColon > 0) {
-      return signature.substring(firstColon + 1);
-    }
-    return signature;
-  }
-
-  private async transformTokenPoolCreationEvent(
-    subName: string,
-    event: TokenPoolCreationEvent,
-  ): Promise<WebSocketMessage | undefined> {
-    const { data: output } = event;
-    const decodedData = decodeHex(output.data ?? '');
-
-    if (event.address.toLowerCase() !== this.service.factoryAddress) {
-      this.logger.warn(`Ignoring token pool creation from unknown address: ${event.address}`);
-      return undefined;
-    }
-
-    const type = output.is_fungible ? TokenType.FUNGIBLE : TokenType.NONFUNGIBLE;
-    const withData = await this.service.supportsData(newContext(), output.contract_address, type);
-    const schema = getTokenSchema(type, withData);
-    const poolLocator: IValidPoolLocator = {
-      address: output.contract_address.toLowerCase(),
-      type,
-      schema,
-    };
-
-    return {
-      event: 'token-pool',
-      data: <TokenPoolEvent>{
-        standard: type === TokenType.FUNGIBLE ? 'ERC20' : 'ERC721',
-        poolLocator: packPoolLocator(poolLocator),
-        type,
-        signer: event.inputSigner,
-        data: decodedData,
-        symbol: output.symbol,
-        info: {
-          name: output.name,
-          address: output.contract_address,
-          schema,
-        },
-        blockchain: {
-          id: this.formatBlockchainEventId(event),
-          name: this.stripParamsFromSignature(this.trimEventSignature(event.signature)),
-          location: 'address=' + event.address,
-          signature: this.trimEventSignature(event.signature),
-          timestamp: event.timestamp,
-          output,
-          info: {
-            blockNumber: event.blockNumber,
-            transactionIndex: event.transactionIndex,
-            transactionHash: event.transactionHash,
-            logIndex: event.logIndex,
-            address: event.address,
-            signature: this.trimEventSignature(event.signature),
-          },
-        },
-      },
-    };
-  }
-
-  private async transformTransferEvent(
-    subName: string,
-    event: TransferEvent,
-  ): Promise<WebSocketMessage | undefined> {
-    const { data: output } = event;
-    const unpackedSub = unpackSubscriptionName(subName);
-    const decodedData = decodeHex(event.inputArgs?.data ?? '');
-
-    if (output.from === ZERO_ADDRESS && output.to === ZERO_ADDRESS) {
-      // should not happen
-      return undefined;
-    }
-    if (unpackedSub.poolLocator === undefined) {
-      // should not happen
-      return undefined;
-    }
-
-    const eventId = this.formatBlockchainEventId(event);
-    const poolLocator = unpackPoolLocator(unpackedSub.poolLocator);
-    const commonData = {
-      id: eventId,
-      poolData: unpackedSub.poolData,
-      poolLocator: unpackedSub.poolLocator,
-      amount: poolLocator.type === TokenType.FUNGIBLE ? output.value : '1',
-      signer: event.inputSigner,
-      data: decodedData,
-      blockchain: {
-        id: eventId,
-        name: this.stripParamsFromSignature(this.trimEventSignature(event.signature)),
-        location: 'address=' + event.address,
-        signature: this.trimEventSignature(event.signature),
-        timestamp: event.timestamp,
-        output,
-        info: {
-          address: event.address,
-          blockNumber: event.blockNumber,
-          transactionIndex: event.transactionIndex,
-          transactionHash: event.transactionHash,
-          logIndex: event.logIndex,
-          signature: this.trimEventSignature(event.signature),
-        },
-      },
-    } as TokenTransferEvent;
-
-    if (poolLocator.type === TokenType.NONFUNGIBLE && output.tokenId !== undefined) {
-      commonData.tokenIndex = output.tokenId;
-      commonData.uri = await this.getTokenUri(
-        newContext(),
-        output.tokenId,
-        poolLocator.address ?? '',
-      );
-    }
-
-    if (output.from === ZERO_ADDRESS) {
-      return {
-        event: 'token-mint',
-        data: { ...commonData, to: output.to } as TokenMintEvent,
-      };
-    } else if (output.to === ZERO_ADDRESS) {
-      return {
-        event: 'token-burn',
-        data: { ...commonData, from: output.from } as TokenBurnEvent,
-      };
-    } else {
-      return {
-        event: 'token-transfer',
-        data: { ...commonData, from: output.from, to: output.to } as TokenTransferEvent,
-      };
-    }
-  }
-
-  private transformApprovalEvent(
-    subName: string,
-    event: ERC20ApprovalEvent | ERC721ApprovalEvent,
-  ): WebSocketMessage | undefined {
-    const { data: output } = event;
-    const unpackedSub = unpackSubscriptionName(subName);
-    const decodedData = decodeHex(event.inputArgs?.data ?? '');
-
-    if (unpackedSub.poolLocator === undefined) {
-      // should not happen
-      return undefined;
-    }
-    const poolLocator = unpackPoolLocator(unpackedSub.poolLocator);
-
-    let operator: string;
-    let subject: string | undefined;
-    let approved = true;
-    if (poolLocator.type === TokenType.FUNGIBLE) {
-      const erc20Event = event as ERC20ApprovalEvent;
-      operator = erc20Event.data.spender;
-      subject = `${output.owner}:${operator}`;
-      approved = BigInt(erc20Event.data.value ?? 0) > BigInt(0);
-    } else {
-      const erc721Event = event as ERC721ApprovalEvent;
-      operator = erc721Event.data.approved;
-      subject = erc721Event.data.tokenId;
-      approved = operator !== ZERO_ADDRESS;
-    }
-
-    const eventId = this.formatBlockchainEventId(event);
-    return {
-      event: 'token-approval',
-      data: <TokenApprovalEvent>{
-        id: eventId,
-        poolData: unpackedSub.poolData,
-        subject,
-        poolLocator: unpackedSub.poolLocator,
-        operator,
-        approved,
-        signer: output.owner,
-        data: decodedData,
-        info: output,
-        blockchain: {
-          id: eventId,
-          name: this.stripParamsFromSignature(this.trimEventSignature(event.signature)),
-          location: 'address=' + event.address,
-          signature: this.trimEventSignature(event.signature),
-          timestamp: event.timestamp,
-          output,
-          info: {
-            blockNumber: event.blockNumber,
-            transactionIndex: event.transactionIndex,
-            transactionHash: event.transactionHash,
-            logIndex: event.logIndex,
-            address: event.address,
-            signature: this.trimEventSignature(event.signature),
-          },
-        },
-      },
-    };
-  }
-
-  private transformApprovalForAllEvent(
-    subName: string,
-    event: ApprovalForAllEvent,
-  ): WebSocketMessage | undefined {
-    const { data: output } = event;
-    const unpackedSub = unpackSubscriptionName(subName);
-    const decodedData = decodeHex(event.inputArgs?.data ?? '');
-
-    if (unpackedSub.poolLocator === undefined) {
-      // should not happen
-      return undefined;
-    }
-
-    const eventId = this.formatBlockchainEventId(event);
-    return {
-      event: 'token-approval',
-      data: <TokenApprovalEvent>{
-        id: eventId,
-        poolData: unpackedSub.poolData,
-        subject: `${output.owner}:${output.operator}`,
-        poolLocator: unpackedSub.poolLocator,
-        operator: output.operator,
-        approved: output.approved,
-        signer: output.owner,
-        data: decodedData,
-        info: output,
-        blockchain: {
-          id: eventId,
-          name: this.stripParamsFromSignature(event.signature),
-          location: 'address=' + event.address,
-          signature: this.trimEventSignature(event.signature),
-          timestamp: event.timestamp,
-          output,
-          info: {
-            blockNumber: event.blockNumber,
-            transactionIndex: event.transactionIndex,
-            transactionHash: event.transactionHash,
-            logIndex: event.logIndex,
-            address: event.address,
-            signature: this.trimEventSignature(event.signature),
-          },
-        },
-      },
-    };
   }
 }
